# Speedup Sinkhorn with einsum + bench