--- a/Makefile
+++ b/Makefile
@@ -57,6 +57,14 @@ rdoc :
 notebook :
 	ipython notebook --matplotlib=inline  --notebook-dir=notebooks/
 	
+bench :
+	echo 'Branch master'
+	git checkout master
+	python3 $(script)
+	echo 'Branch $(branch)'
+	git checkout $(branch)
+	python3 $(script)
+	
 autopep8 :
 	autopep8 -ir test ot examples --jobs -1