--- a/Makefile
+++ b/Makefile
@@ -59,12 +59,12 @@ notebook :
 	ipython notebook --matplotlib=inline  --notebook-dir=notebooks/
 	
 bench :
-	echo 'Branch master'
+	@echo 'Branch master'
 	git checkout master
-	python3 $(script)
-	echo 'Branch $(branch)'
+	#python3 $(script)
+	@echo 'Branch $(branch)'
 	git checkout $(branch)
-	python3 $(script)
+	#python3 $(script)
 	
 autopep8 :
 	autopep8 -ir test ot examples --jobs -1