--- a/ot/bregman.py
+++ b/ot/bregman.py
@@ -358,9 +358,9 @@ def sinkhorn_knopp(a, b, M, reg, numItermax=1000,
     while (err > stopThr and cpt < numItermax):
         uprev = u
         vprev = v
-        KtransposeU = np.dot(K.T, u)
+        KtransposeU = np.einsum('ij,i->j',K,u)#np.dot(K.T, u)
         v = np.divide(b, KtransposeU)
-        u = 1. / np.dot(Kp, v)
+        u = 1. / np.einsum('ij,j->i',Kp,v)#np.dot(Kp, v)
 
         if (np.any(KtransposeU == 0) or
                 np.any(np.isnan(u)) or np.any(np.isnan(v)) or